imp: move TmValidation/ExecutionContext into ICS-02 for general usage
