+++ ibc-clients/ics07-tendermint/src/client_state/update_client.rs
@@ -1,5 +1,6 @@
 use ibc_client_tendermint_types::error::{Error, IntoResult};
 use ibc_client_tendermint_types::{ConsensusState as ConsensusStateType, Header as TmHeader};
+use ibc_core_client::context::{ExtClientValidationContext, TypeCaster};
 use ibc_core_client::types::error::ClientError;
 use ibc_core_client::types::Height;
 use ibc_core_host::types::identifiers::{ChainId, ClientId};
@@ -11,9 +12,7 @@ use tendermint_light_client_verifier::options::Options;
 use tendermint_light_client_verifier::types::{TrustedBlockState, UntrustedBlockState};
 use tendermint_light_client_verifier::Verifier;
 
-use crate::context::{
-    ConsensusStateConverter, TmVerifier, ValidationContext as TmValidationContext,
-};
+use crate::context::TmVerifier;
 
 pub fn verify_header<V, H>(
     ctx: &V,
@@ -24,8 +23,8 @@ pub fn verify_header<V, H>(
     verifier: &impl TmVerifier,
 ) -> Result<(), ClientError>
 where
-    V: TmValidationContext,
-    V::ConsensusStateRef: ConsensusStateConverter,
+    V: ExtClientValidationContext,
+    V::ConsensusStateRef: TypeCaster<ConsensusStateType, ClientError>,
     H: MerkleHash + Sha256 + Default,
 {
     // Checks that the header fields are valid.
@@ -110,8 +109,8 @@ pub fn check_for_misbehaviour_on_update<V>(
     client_latest_height: &Height,
 ) -> Result<bool, ClientError>
 where
-    V: TmValidationContext,
-    V::ConsensusStateRef: ConsensusStateConverter,
+    V: ExtClientValidationContext,
+    V::ConsensusStateRef: TypeCaster<ConsensusStateType, ClientError>,
 {
     let maybe_existing_consensus_state = {
         let path_at_header_height = ClientConsensusStatePath::new(
